Match fn_800A1040 by dropping redundant (s16) cast
p1 is already s16; the explicit (s16)p1 cast in the loop comparison
forced MWCC to hoist the sign-extension into a saved reg. Dropping it
lets the value re-extend at the use site like the target. 87.4% -> 100%.

(cherry picked from commit 1f85632908b9f2e084d160ba46776493de73c02b)

+++ src/main/dll/dll_000B_dll0b.c
@@ -1564,7 +1564,7 @@ void fn_800A1040(s16 p1, int p2)
     for (i = 0; i < PARTFX_ACTIVE_EFFECT_COUNT; i++)
     {
         if (arr[i] == NULL) continue;
-        if ((s16)p1 != arr[i]->sequenceId && p2 == 0) continue;
+        if (p1 != arr[i]->sequenceId && p2 == 0) continue;
         if (arr[i]->auxAllocation != NULL)
         {
             mm_free(arr[i]->auxAllocation);
